backend/requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.34.0
google-generativeai==0.8.3
firebase-admin==6.7.0
httpx==0.28.1
python-dotenv==1.1.0
pydantic==2.11.2
google-auth==2.38.0
google-cloud-logging>=3.9.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.14.0
